tradeguard/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
supabase==2.11.0
anthropic==0.42.0
pydantic==2.10.4
python-multipart==0.0.20
python-dateutil==2.9.0
